Fix compile err

+++ examples/desktop/utils/helper.h
@@ -21,7 +21,7 @@ int selectCamera(CcapProvider* provider);
 void createDirectory(const char* path);
 // Get current working directory (portable)
 // Returns 0 on success, -1 on failure
-int getCurrentWorkingDirectory(char* buffer, size_t size);
+int getCurrentWorkingDirectory(char* buffer, int size);
 
 #ifdef __cplusplus
 } // extern "C"
